Move UpdateButton from MenuBar to its own folder
