+++ packages/app/src/components/VisualizationOptions/VisualizationOptionsManager.js
@@ -6,7 +6,7 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import i18n from '@dhis2/d2-i18n';
 
-import UpdateButton from '../MenuBar/UpdateButton';
+import UpdateButton from '../UpdateButton/UpdateButton';
 import VisualizationOptions from './VisualizationOptions';
 
 class VisualizationOptionsManager extends Component {
